add ara

+++ hh_mtusi/hhapp/models.py
@@ -5,7 +5,7 @@ class Vacancy(models.Model):
     name = models.CharField(max_length=255)  # Название вакансии
     company = models.CharField(max_length=255)  # Название компании
     salary = models.DecimalField(max_digits=10, decimal_places=2, default=1.00 ,null=True)
-    area = models.DecimalField(max_digits=10, decimal_places=2, default=1.00 ,null=True)
+    area =  models.IntegerField(default=1,null=True) 
     url = models.URLField()  # URL вакансии
     published_at = models.DateTimeField()  # Дата публикации
